refactor: move Checksums to internals
