+++ clients/src/main/java/org/apache/kafka/common/record/internal/LegacyRecord.java
@@ -20,7 +20,7 @@ import org.apache.kafka.common.KafkaException;
 import org.apache.kafka.common.errors.CorruptRecordException;
 import org.apache.kafka.common.record.TimestampType;
 import org.apache.kafka.common.utils.ByteBufferOutputStream;
-import org.apache.kafka.common.utils.Checksums;
+import org.apache.kafka.common.utils.internals.Checksums;
 import org.apache.kafka.common.utils.Utils;
 import org.apache.kafka.common.utils.internals.ByteUtils;
