+++ clients/src/main/java/org/apache/kafka/common/utils/internals/Checksums.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.common.utils;
+package org.apache.kafka.common.utils.internals;
 
 import java.nio.ByteBuffer;
 import java.util.zip.Checksum;
