+++ clients/src/main/java/org/apache/kafka/common/utils/internals/Crc32C.java
@@ -17,8 +17,6 @@
 
 package org.apache.kafka.common.utils.internals;
 
-import org.apache.kafka.common.utils.Checksums;
-
 import java.nio.ByteBuffer;
 import java.util.zip.CRC32C;
 import java.util.zip.Checksum;
